LPD-91092 Split the data masking tests by the class under test
